# Improve performance of getSizeUnit function