--- a/Sizy/Program.fs
+++ b/Sizy/Program.fs
@@ -3,6 +3,7 @@ module Sizy.Program
 open Sizy.Config
 
 open System
+open System.IO
 open System.IO.Abstractions
 open System.Collections.Generic
 open System.Collections.Concurrent
@@ -12,18 +13,18 @@ let SizeUnits = [ "B"; "k"; "M"; "G"; "T"; "P"; "E" ]
 
 type Entry(path: string, size: int64, isDir: bool, sep: char) =
 
-    member this.name =
+    member this.Name =
         Array.last (path.Split sep) + if isDir then string sep else ""
 
-    member this.size = size
-    member this.isDir = isDir
+    member this.Size = size
+    member this.IsDir = isDir
 
 let rec getSize (fs: IFileSystem) (fsEntries: IDictionary<_, _>) (errors: IDictionary<_, _>) path =
     try
         let attr = fs.File.GetAttributes path
 
         let size, isDir =
-            if attr.HasFlag System.IO.FileAttributes.Directory
+            if attr.HasFlag FileAttributes.Directory
             then fs.Directory.EnumerateFileSystemEntries path |> Seq.sumBy (getSize fs fsEntries errors), true
             else fs.FileInfo.FromFileName(path).Length, false
         fsEntries.[path] <- Entry(path, size, isDir, fs.Path.DirectorySeparatorChar)
@@ -39,15 +40,6 @@ let sizyMain (fs: IFileSystem, path: string) =
     let sizes = PSeq.map (getSize fs fsEntries errors) ls
     let totSize = PSeq.sum sizes
     ls, fsEntries, totSize, errors
-
-let getSizeUnit2 bytes =
-    if bytes <= 0L then
-        0.0, SizeUnits.[0]
-    else
-        let bytesF = float (bytes)
-        let sizeUnitsIdx = Math.Floor(Math.Log(bytesF, 1024.0))
-        let num = Math.Round(bytesF / Math.Pow(1024.0, sizeUnitsIdx), 0)
-        num, SizeUnits.[int (sizeUnitsIdx)]
 
 let getSizeUnit bytes =
     if bytes <= 0L then
@@ -73,9 +65,9 @@ let getSizeString name size =
 
 [<EntryPoint>]
 let main argv =
-    match Config.getConfiguration argv with
+    match getConfiguration argv with
     | Config config ->
-        let fs = new FileSystem()
+        let fs = FileSystem()
 
         let path =
             if config.Contains InputPath then config.GetResult InputPath else fs.Directory.GetCurrentDirectory()
@@ -86,9 +78,9 @@ let main argv =
         let print f =
             PSeq.filter f ls
             |> PSeq.sort
-            |> Seq.iter (fun p -> printf "%s" (getSizeString fsEntries.[p].name fsEntries.[p].size))
-        print (fun x -> fsEntries.ContainsKey x && fsEntries.[x].isDir)
-        print (fun x -> fsEntries.ContainsKey x && not fsEntries.[x].isDir)
+            |> Seq.iter (fun p -> printf "%s" (getSizeString fsEntries.[p].Name fsEntries.[p].Size))
+        print (fun x -> fsEntries.ContainsKey x && fsEntries.[x].IsDir)
+        print (fun x -> fsEntries.ContainsKey x && not fsEntries.[x].IsDir)
 
         let totSize, totSizeUnit = getSizeUnit totSize
         printfn "%s\n%10.0f %-1s" (String.replicate 12 "-") totSize totSizeUnit