The new `getSizeUnit` function uses bit shift operations instead of Math.Pow, Math.Log, etc, and appears to be up to almost 4x faster.

Credits: https://www.somacon.com/p576.php

``sudo dotnet run -c Release --framework netcoreapp3.1``

``` ini

BenchmarkDotNet=v0.12.0, OS=arch 
Intel Core i5-7200U CPU 2.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.1.103
  [Host]     : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 3.1.3 (CoreCLR 4.700.20.11803, CoreFX 4.700.20.12001), X64 RyuJIT


```
|       Method | NBytes |     Mean |    Error |    StdDev |   Median |
|------------- |------- |---------:|---------:|----------:|---------:|
|  **GetSizeUnit** |      **0** | **14.79 ns** | **0.332 ns** |  **0.828 ns** | **14.51 ns** |
| GetSizeUnit2 |      0 | 18.22 ns | 0.900 ns |  2.596 ns | 17.64 ns |
|  **GetSizeUnit** |  **10010** | **27.27 ns** | **1.045 ns** |  **3.033 ns** | **26.53 ns** |
| GetSizeUnit2 |  10010 | 86.79 ns | 1.742 ns |  3.052 ns | 86.90 ns |
|  **GetSizeUnit** |  **20020** | **26.60 ns** | **1.272 ns** |  **3.610 ns** | **25.98 ns** |
| GetSizeUnit2 |  20020 | 99.43 ns | 5.018 ns | 14.796 ns | 95.34 ns |
|  **GetSizeUnit** |  **30030** | **30.99 ns** | **1.970 ns** |  **5.810 ns** | **29.17 ns** |
| GetSizeUnit2 |  30030 | 97.80 ns | 3.663 ns | 10.627 ns | 94.70 ns |
|  **GetSizeUnit** |  **40040** | **28.73 ns** | **1.992 ns** |  **5.872 ns** | **26.28 ns** |
| GetSizeUnit2 |  40040 | 86.64 ns | 1.754 ns |  2.625 ns | 88.17 ns |
|  **GetSizeUnit** |  **50050** | **26.23 ns** | **0.781 ns** |  **2.240 ns** | **25.17 ns** |
| GetSizeUnit2 |  50050 | 89.46 ns | 1.825 ns |  4.083 ns | 89.84 ns |
|  **GetSizeUnit** |  **60060** | **26.45 ns** | **0.823 ns** |  **2.336 ns** | **25.25 ns** |
| GetSizeUnit2 |  60060 | 86.02 ns | 1.756 ns |  2.885 ns | 85.15 ns |
|  **GetSizeUnit** |  **70070** | **22.63 ns** | **1.090 ns** |  **3.213 ns** | **20.87 ns** |
| GetSizeUnit2 |  70070 | 86.58 ns | 1.774 ns |  2.864 ns | 86.20 ns |
|  **GetSizeUnit** |  **80080** | **26.10 ns** | **0.930 ns** |  **1.579 ns** | **25.95 ns** |
| GetSizeUnit2 |  80080 | 89.28 ns | 1.830 ns |  4.522 ns | 89.07 ns |
|  **GetSizeUnit** |  **90090** | **31.68 ns** | **1.889 ns** |  **5.569 ns** | **30.56 ns** |
| GetSizeUnit2 |  90090 | 86.25 ns | 1.751 ns |  3.158 ns | 85.64 ns |